requirements.prod.txt
flask
flask-login
flask-wtf
flask-session
pyodbc
SQLAlchemy
wtforms
email-validator
jinja2
docxtpl
python-dateutil
python-docx
openpyxl
requests
wechatpy
passlib
bcrypt
cryptography
python-dotenv
gunicorn
